RD/requirements.docker.txt
fastapi
uvicorn[standard]
itsdangerous
python-multipart
pillow
openai
opencv-python-headless
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
